refactor of app and decorators
