+++ api/bare_app.py
@@ -1,0 +1,4 @@
+from flask import Flask
+
+app = Flask(__name__)
+app.config.from_object('config.Config')
