+++ api/v1/resources/harvest.py
@@ -1,12 +1,14 @@
-import os
 import json
 import urllib2
 from urlparse import urlparse
-from flask import jsonify
-from flask_restful import Resource, reqparse
+
 import ckanapi
 from tasks import harvesting
-
+import os
+from api.auth.auth import auth
+from ckanops import dcat_to_utf8_dict, munge, converters, upsert_dataset
+from flask import jsonify
+from flask_restful import Resource, reqparse
 
 HOST = os.getenv('CKAN_HOST')
 TOKEN = os.getenv('CKAN_API_TOKEN')
@@ -16,6 +18,7 @@ remote = ckanapi.RemoteCKAN(HOST, user_agent='ckanops/1.0', apikey=TOKEN)
 
 
 class Harvest(Resource):
+    decorators=[auth.login_required]
     def post(self):
         parser = reqparse.RequestParser(bundle_errors=True)
         parser.add_argument('url', type=str, required=True)
